django_admin/requirements.txt
aiocontextvars==0.2.2
attrs==19.3.0
certifi==2019.11.28
chardet==3.0.4
contextvars==2.4
Django==2.2.7
django-adminlte3==0.1.2
django-model-utils==4.0.0
django-notifications-hq==1.5.0
idna==2.8
idna-ssl==1.1.0
immutables==0.11
gunicorn==20.0.4
jsonfield==2.0.2
loguru==0.4.0
multidict==4.7.2
psycopg2==2.8.4
pytz==2019.3
requests==2.22.0
sqlparse==0.3.0
typing-extensions==3.7.4.1
urllib3==1.25.7
yarl==1.4.2
